Revert "change names, add monitoring feature"
